+++ src/lib.rs
@@ -77,6 +77,7 @@ pub mod user;
 pub mod entities;
 pub mod cursor;
 pub mod tweet;
+pub mod search;
 mod links;
 mod common;